% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_airbnb_to_postgres.R
\name{build_pgdb}
\alias{build_pgdb}
\title{Build a PostgreSQL database from the tables hosted by insideairbnb.}
\usage{
build_pgdb(con, cities, table_names)
}
\arguments{
\item{con}{A database connection object.}

\item{cities}{A list of cities to take tables from. All cities must
be hosted by insideairbnb.}

\item{table_names}{A list of tables to be built. Can be 'calendar', 
'detailed_listings', 'summary_listings', detailed_review',
'summary_review', neighbourhood', or 'neighbourhood_geo'.}
}
\value{
A list of the number of rows affected by the SQL query for
each URL.
}
\description{
Build a PostgreSQL database from the tables hosted by insideairbnb.
}
\examples{
library(DBI)
con <- DBI::dbConnect(odbc::odbc(), Driver = "PostgreSQL ODBC Driver(ANSI)", 
                      Server = "localhost", Database = "inside_airbnb", 
                      UID = rstudioapi::askForPassword("Database user"), 
                      PWD = rstudioapi::askForPassword("Database password"), 
                      Port = 5432)
tmp <- dbExecute(con, "SET CLIENT_ENCODING TO 'utf8'")
url_table <- scrape_inside_airbnb()
tables <- c('calendar', 'detailed_listings')
city_names <- c('San Francisco', 'Portland')
build_pgb(con, cities = city_names, table_names = tables)
}
