% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_airbnb_to_postgres.R
\name{create_table}
\alias{create_table}
\title{Create an empty SQL table for one of the valid table types}
\usage{
create_table(con, table_name)
}
\arguments{
\item{con}{A database connection object}

\item{table_name}{Charactor specifying the table to be created. Can be
'calendar', 'detailed_listings', 'summary_listings', detailed_review',
'summary_review', neighbourhood', or 'neighbourhood_geo'.}
}
\value{
TRUE upon completion
}
\description{
Create an empty SQL table for one of the valid table types
}
\examples{
library(DBI)
con <- DBI::dbConnect(odbc::odbc(), Driver = "PostgreSQL ODBC Driver(ANSI)", 
                      Server = "localhost", Database = "inside_airbnb", 
                      UID = rstudioapi::askForPassword("Database user"), 
                      PWD = rstudioapi::askForPassword("Database password"), 
                      Port = 5432)
tmp <- dbExecute(con, "SET CLIENT_ENCODING TO 'utf8'")
create_table(con, table_name = 'calendar')
}
