% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{drop_junk}
\alias{drop_junk}
\title{Remove junk columns from a data frame.}
\usage{
drop_junk(df, cols = junk_cols)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{A list of column names to be dropped 
from the data frame.}
}
\value{
A data frame with junk columns dropped.
}
\description{
Drops columns with mostly missing data or only a
single distinct value. Allows user to provide a 
list of additional columns to drop.
}
\examples{
junk_df <- data.frame(a = rep('a', times = 10),
                      b = 1:10,
                      c = c(rep(NA, times = 6),
                            1:4),
                      d = 3:13)
clean_df <- drop_junk(junk_df, cols = c('d'))
head(clean_df)
}
