% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{factorize_amenities}
\alias{factorize_amenities}
\title{Add factor columns indicating listing amenities.}
\usage{
factorize_amenities(df, amenities = amenities_list, prop_filter = 0.1)
}
\arguments{
\item{df}{A data frame like 'detailed_listings' with an amenities 
column where each record contains a list of amenities the listing has.}

\item{amenities}{A list of amenities to detect and convert to factors.}

\item{prop_filter}{The minimum poportion of either TRUE or FALSE 
required for an amenity to be included.}
}
\value{
A data frame with indicator columns for each amenity.
}
\description{
Add factor columns indicating listing amenities.
}
\examples{
dirty_df <- data.frame(amenities = c({tv,lamp},{tv},{lamp}))
am_list <- c('tv', 'lamp')
clean_df <- factorize_amenities(dirty_df, amenities = am_list)
}
