% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{factorize_booleans}
\alias{factorize_booleans}
\title{Detect and convert boolean variables to factors.}
\usage{
factorize_booleans(df, bools = bool_chars)
}
\arguments{
\item{df}{A data frame with some columns containing only 
two distinct characters.}

\item{bools}{A list of two characters to detect.}
}
\value{
A data frame with the boolean columns converted 
to factors.
}
\description{
Detect and convert boolean variables to factors.
}
\examples{
char_df <- data.frame(bool_col = rep(c('t','f'), times = 5))
factor_df <- convert_dates(char_df, c('t', 'f'))
is.factor(date_df$bool_col)
}
