% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{factorize_cancellation_policy}
\alias{factorize_cancellation_policy}
\title{Convert cancellation policy to 'flexible', 'moderate' or 'strict' and factorize.}
\usage{
factorize_cancellation_policy(df)
}
\arguments{
\item{df}{A data frame with a 'cancellation_policy' column where entries have
different codes that all contain aither 'flexible', 'moderate', or 'strict'.}
}
\value{
A data frame with the 'cancellation_policy' column recoded and converted 
to factor.
}
\description{
Convert cancellation policy to 'flexible', 'moderate' or 'strict' and factorize.
}
\examples{
can_df <- data.frame(cancellation_policy = c('moderate','kinda strict', 'mostly flexible'))
fac_df <- factorize_cancellation_policy(can_df)
is.factor(fac_df$cancellation_policy)
levels(fac_df$cancellation_policy)
}
