% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{factorize_license}
\alias{factorize_license}
\title{Add a column to indicate if a host is licensed.}
\usage{
factorize_license(df)
}
\arguments{
\item{df}{A data frame with a column of license types where NA
denotes the absense of a license.}
}
\value{
A data frame with a 'host_has_license' column where 
not NA is TRUE and NA is FALSE.
}
\description{
Add a column to indicate if a host is licensed.
}
\examples{
lic_df <- data.frame(license = c('sfgert34',NA,'ST$\%'))
fac_df <- factorize_license(lic_df)
is.factor(fac_df$host_has_license)
}
