% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{factorize_simple_cols}
\alias{factorize_simple_cols}
\title{Convert columns with simple categorical variables to factors.}
\usage{
factorize_simple_cols(df, factor_cols = simple_factor_list)
}
\arguments{
\item{df}{A data frame with some categorical columns that need to be factors.}

\item{factor_cols}{List of column names to covert.}
}
\value{
A data frame with the elected columns converted to factors.
}
\description{
Convert columns with simple categorical variables to factors.
}
\examples{
char_df <- data.frame(fac_col = rep(c("level_one","level_two", "level_three"), 
                                    each = 5))
fac_df <- factorize_simple_cols(char_df, factor_cols = c('fac_col'))
is.factor(fac_df$fac_col)
}
