% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{host_rates_to_numeric}
\alias{host_rates_to_numeric}
\title{Convert host percentage rates columns to numeric.}
\usage{
host_rates_to_numeric(df, rate_cols = rate_list)
}
\arguments{
\item{df}{A data frame with character columns containing percentage values.}

\item{rate_cols}{A list of column names of the percentage columns.}
}
\value{
A data frame with the percent symbols reoved and converted to numeric.
}
\description{
Convert host percentage rates columns to numeric.
}
\examples{
perc_df <- data.frame(rates <- c('10\%', '2\%', '100\%'))
num_df <- host_rates_to_numeric(perc_df, rate_cols = c('rates'))
is.numeric(num_df$rates)
}
