% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{process_missing}
\alias{process_missing}
\title{Process columns with missing values depending on column name.}
\usage{
process_missing(df, proportion_to_drop = 0.05)
}
\arguments{
\item{df}{A data frame like 'detailed_listings' that has columns with
missing values that need to be dealt with in specialized ways.
Supported column names are: 'host_response_rate', host_acceptance_rate',
'first_review', 'last_review', and 'review_scores_*'.}

\item{proportion_to_drop}{If a column has less than this proportion
of missing records, then missings records will be deleted.}
}
\value{
A data frame with processed missing values.
}
\description{
Process columns with missing values depending on column name.
}
\examples{
To-do
}
