% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{refactor_bed_type}
\alias{refactor_bed_type}
\title{Refactor bed types to either 'Real Bed' or 'Fake Bed'}
\usage{
refactor_bed_type(df)
}
\arguments{
\item{df}{A data frame with a 'bed_type' column where many of the
entries are already 'Real Bed', but a few are random types of 
other beds.}
}
\value{
A data frame where the 'Real Bed' entries stay the same, 
and all other are converted to 'Fake Bed' and factorized.
}
\description{
Refactor bed types to either 'Real Bed' or 'Fake Bed'
}
\examples{
bed_df <- data.frame(bed_type = c('Real Bed', 'hammock', 'couch'))
fac_df <- refactor_bed_type(bed_df)
is.factor(fac_df$bed_type)
levels(fac_df$bed_type)
}
