% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_airbnb_to_postgres.R
\name{scrape_inside_airbnb}
\alias{scrape_inside_airbnb}
\title{Scrape the data URLs from insideairbnb.com}
\source{
\url{http://insideairbnb.com/get-the-data.html}
}
\usage{
scrape_inside_airbnb()
}
\value{
A data frame with 11 variables:
\describe{
  \item{url_country}{country for the URL's}
  \item{url_state}{state or province for the URL's}
  \item{url_city}{state or province for the URL's}
  \item{url_date}{date the data was scraped from Airbnb}
  \item{detailed_listings_url}{URL to listings.csv.gz}
  \item{summary_listings_url}{URL to listings.csv}
  \item{calendar_url}{URL to calendar.csv.gz}
  \item{detailed_review_url}{URL to reviews.csv.gz}
  \item{summary_review_url}{URL to reviews.csv}
  \item{neighbourhood_url}{URL to neighbourhoods.csv}
  \item{neighbourhood_geo_url}{URL to neighourhoods.geojson}
}
}
\description{
Scrape the data URLs from insideairbnb.com
}
