% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utils.R
\name{text_to_numeric}
\alias{text_to_numeric}
\title{Convert columns with text entries to numeric columns given by the length of the text.}
\usage{
text_to_numeric(df, cols = text_cols)
}
\arguments{
\item{df}{A data frame with some columns where each record is a text description.}

\item{cols}{A list of column names specifying the text columns.}
}
\value{
A data frame with text columns converted to numeric.
}
\description{
Convert columns with text entries to numeric columns given by the length of the text.
}
\examples{
text_df = data.frame(text_col = c('short description', 'not very long description'))
num_df <- text_to_numeric(text_df, cols = c('text_col'))
is.numeric(num_df$text_col)
}
