% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_airbnb_to_postgres.R
\name{url_to_pgdb}
\alias{url_to_pgdb}
\title{Load data from a URL to a PostrgreSQL database.}
\usage{
url_to_pgdb(con, city, table_name, url_df)
}
\arguments{
\item{con}{A database connection object}

\item{city}{Character specifying a city name. Must be a city hosted by
insideairbnb.}

\item{table_name}{Charactor specifying the table to be created. Can be
'calendar', 'detailed_listings', 'summary_listings', detailed_review',
'summary_review', neighbourhood', or 'neighbourhood_geo'.}

\item{url_df}{A data frame as returned by \code{\link{scrape_inside_airbnb}}}
}
\value{
Number of rows affected by corresponding SQL query.
}
\description{
Loads the most recently scraped URL corresponding to the supplied
cities and tables.
}
\examples{
library(DBI)
con <- DBI::dbConnect(odbc::odbc(), Driver = "PostgreSQL ODBC Driver(ANSI)", 
                      Server = "localhost", Database = "inside_airbnb", 
                      UID = rstudioapi::askForPassword("Database user"), 
                      PWD = rstudioapi::askForPassword("Database password"), 
                      Port = 5432)
tmp <- dbExecute(con, "SET CLIENT_ENCODING TO 'utf8'")
url_table <- scrape_inside_airbnb()
create_table(con, table_name = 'calendar')
url_to_pgb(con, city = 'San Francisco', table_name = 'calendar', url_df = url_table)
}
